requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
supabase==2.15.0
openai==1.107.3
openai-agents==0.3.0
numpy==1.26.4
beautifulsoup4==4.13.4
pandas==2.0.3
requests==2.32.3
python-multipart==0.0.9
psycopg2-binary==2.9.9
pinecone==6.0.2
python-dotenv==1.0.1
reportlab==4.4.0
aiohttp==3.11.16
redis==5.0.1
celery==5.5.2
PyPDF2
docx
sendgrid==6.10.0
docxtpl
python-docx
docxcompose
docx2pdf
stripe
email-validator==2.3.0
playwright==1.55.0
azure-identity
azure-servicebus
azure-storage-blob
openpyxl